Refine mounted action schema cache cleanup
Amp-Thread-ID: https://ampcode.com/threads/T-01a04e2a-fe3c-731e-b615-9a4456fdceaa
