+++ tests/src/Actions/ActionTest.php
@@ -6,7 +6,6 @@ use Filament\Actions\Concerns\InteractsWithActions;
 use Filament\Actions\Contracts\HasActions;
 use Filament\Actions\Enums\ActionStatus;
 use Filament\Actions\Testing\TestAction;
-use Filament\Forms\Components\TextInput;
 use Filament\Notifications\Notification;
 use Filament\Schemas\Concerns\InteractsWithSchemas;
 use Filament\Schemas\Contracts\HasSchemas;
@@ -479,6 +478,48 @@ describe('nested actions', function (): void {
                 ]),
             ])
             ->assertDispatched('arguments-test-called', foo: $foo, bar: $bar, baz: $baz);
+    });
+});
+
+describe('unmounting actions', function (): void {
+    it('forgets a nested action schema before another action is mounted at the same index', function (): void {
+        $livewire = livewire(Actions::class)
+            ->mountAction('staleSchemaParent')
+            ->mountAction('first');
+
+        expect(array_keys($livewire->instance()->getSchema('mountedActionSchema1')->getFlatFields()))
+            ->toBe(['alpha']);
+
+        $livewire->call('unmountThenMountAction', 'second');
+
+        expect(array_keys($livewire->instance()->getSchema('mountedActionSchema1')->getFlatFields()))
+            ->toBe(['beta']);
+    });
+
+    it('forgets a root action schema before another action is mounted at the same index', function (): void {
+        $livewire = livewire(Actions::class)
+            ->mountAction('staleSchemaParent');
+
+        expect(array_keys($livewire->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
+            ->toBe(['parentField']);
+
+        $livewire->call('unmountThenMountAction', 'staleSchemaOther');
+
+        expect(array_keys($livewire->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
+            ->toBe(['otherField']);
+    });
+
+    it('preserves the schema of an action that remains mounted', function (): void {
+        $livewire = livewire(Actions::class)
+            ->mountAction('staleSchemaParent')
+            ->mountAction('first');
+
+        $livewire->call('unmountAction');
+
+        expect(array_keys($livewire->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
+            ->toBe(['parentField'])
+            ->and($livewire->instance()->getSchema('mountedActionSchema1'))
+            ->toBeNull();
     });
 });
 
@@ -2381,92 +2422,3 @@ describe('authorization', function (): void {
         });
     });
 });
-
-describe('unmounting actions', function (): void {
-    it('does not hand a closed action\'s schema to the next action mounted in its place', function (): void {
-        $component = livewire(StaleMountedActionSchemaHarness::class)
-            ->mountAction('parent')
-            ->mountAction('first');
-
-        expect(array_keys($component->instance()->getSchema('mountedActionSchema1')->getFlatFields()))
-            ->toBe(['alpha']);
-
-        $component->call('unmountThenMountAction', 'second');
-
-        expect(array_keys($component->instance()->getSchema('mountedActionSchema1')->getFlatFields()))
-            ->toBe(['beta']);
-    });
-
-    it('does not hand it over at the root of the stack either', function (): void {
-        $component = livewire(StaleMountedActionSchemaHarness::class)
-            ->mountAction('parent');
-
-        expect(array_keys($component->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
-            ->toBe(['parentField']);
-
-        $component->call('unmountThenMountAction', 'other');
-
-        expect(array_keys($component->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
-            ->toBe(['otherField']);
-    });
-
-    it('keeps the schema of an action that is still mounted', function (): void {
-        $component = livewire(StaleMountedActionSchemaHarness::class)
-            ->mountAction('parent')
-            ->mountAction('first');
-
-        $component->call('unmountAction');
-
-        expect(array_keys($component->instance()->getSchema('mountedActionSchema0')->getFlatFields()))
-            ->toBe(['parentField'])
-            ->and($component->instance()->getSchema('mountedActionSchema1'))
-            ->toBeNull();
-    });
-});
-
-class StaleMountedActionSchemaHarness extends Component implements HasActions, HasSchemas
-{
-    use InteractsWithActions;
-    use InteractsWithSchemas;
-
-    public function parentAction(): Action
-    {
-        return Action::make('parent')
-            ->schema([
-                TextInput::make('parentField'),
-            ])
-            ->registerModalActions([
-                Action::make('first')
-                    ->schema([
-                        TextInput::make('alpha'),
-                    ])
-                    ->action(fn () => null),
-                Action::make('second')
-                    ->schema([
-                        TextInput::make('beta'),
-                    ])
-                    ->action(fn () => null),
-            ])
-            ->action(fn () => null);
-    }
-
-    public function unmountThenMountAction(string $name): void
-    {
-        $this->unmountAction();
-        $this->mountAction($name);
-    }
-
-    public function otherAction(): Action
-    {
-        return Action::make('other')
-            ->schema([
-                TextInput::make('otherField'),
-            ])
-            ->action(fn () => null);
-    }
-
-    public function render(): View
-    {
-        return view('livewire.form');
-    }
-}
